splitting out some of the graphql types into their own files
