+++ Peer.GitHub/GraphQL/PullRequestSearch/CheckConclusionState.cs
@@ -1,0 +1,16 @@
+﻿namespace Peer.GitHub.GraphQL.PullRequestSearch
+{
+    public enum CheckConclusionState
+    {
+        Action_Required,
+        Cancelled,
+        Failure,
+        Neutral,
+        Skipped,
+        Stale,
+        Startup_Failure,
+        Success,
+        Timed_Out
+    }
+#nullable enable
+}
