+++ Peer.GitHub/GraphQL/PullRequestSearch/CheckRunStatusState.cs
@@ -1,0 +1,13 @@
+﻿namespace Peer.GitHub.GraphQL.PullRequestSearch
+{
+    public enum CheckRunStatusState
+    {
+        Queued,
+        In_Progress,
+        Completed,
+        Waiting,
+        Pending,
+        Requested
+    }
+#nullable enable
+}
