+++ Peer.GitHub/GraphQL/PullRequestSearch/CheckSuite.cs
@@ -1,0 +1,8 @@
+﻿namespace Peer.GitHub.GraphQL.PullRequestSearch
+{
+    public class CheckSuite
+    {
+        public NodeList<CheckRun> CheckRuns { get; set; }
+    }
+#nullable enable
+}
