+++ Peer.GitHub/GraphQL/PullRequestSearch/Commit.cs
@@ -6,10 +6,5 @@
         public StatusCheckRollup StatusCheckRollup { get; set; }
         public NodeList<CheckSuite> CheckSuites { get; set; }
     }
-
-    public class CheckSuite
-    {
-        public NodeList<CheckRun> CheckRuns { get; set; }
-    }
 #nullable enable
 }
