+++ Peer.GitHub/GraphQL/PullRequestSearch/NodeList.cs
@@ -1,0 +1,10 @@
+﻿using System.Collections.Generic;
+
+namespace Peer.GitHub.GraphQL.PullRequestSearch
+{
+    public class NodeList<T>
+    {
+        public List<T> Nodes { get; set; }
+    }
+#nullable enable
+}
