+++ Peer/Properties/launchSettings.json
@@ -2,7 +2,7 @@
   "profiles": {
     "Peer": {
       "commandName": "Project",
-      "commandLineArgs": "details 345"
+      "commandLineArgs": "details 123"
     }
   }
 }
